services/messenger/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.12.4
pydantic-settings==2.11.0

# HTTP Client
httpx==0.28.1

# JWT Authentication
PyJWT==2.10.1

# Database (async)
asyncpg==0.31.0
alembic==1.14.1
sqlalchemy[asyncio]==2.0.41

# Redis (caching, Socket.IO adapter, dramatiq)
redis[hiredis]==5.2.1

# Socket.IO
python-socketio==5.12.1

# Admin panel (per-service)
sqladmin==0.20.1
itsdangerous==2.2.0

# Background workers
dramatiq[redis,watch]==1.17.1

# Observability
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
testcontainers[postgres]==4.9.0
